+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/util/Utilities.java
@@ -199,10 +199,9 @@ public class Utilities {
         }
         if (exitCode != 0) {
             throw new RuntimeException("Process failed with exit code " + exitCode);
-        } else {
-            @SuppressWarnings("unused")
-            boolean success = out.delete();
         }
+        @SuppressWarnings("unused")
+        boolean success = out.delete();
     }
 
     public static void runProcess(String directory, String... commands) throws IOException, InterruptedException {